aider: Added elapsed time to error response in server.py
# Aider chat conversation:

USER: last thing we did was to add "         logsHtml += '</pre>';
         if (!data.path.length) {
             logsHtml += '<p>' + data.time + '</p>';
         }" to main.js but I cannot see the elapsed time as a user.

ASSISTANT: The elapsed time is being added to the logs in the `server/crawler.py` file, but it seems like it's not being displayed on the client side because the `data.time` property is not being set in the server response when an error occurs.

To fix this, we need to modify the `server/server.py` file to include the elapsed time in the error response.

Here is the *SEARCH/REPLACE* block:

server/server.py
```python
<<<<<<< SEARCH
        if not path:
            return jsonify({'error': logs[-1], 'logs': logs, 'time': logs[-1]}), 500
        else:
            return jsonify({'error': 'An error occurred while finding path', 'logs': logs}), 500
=======
        elapsed_time = logs[-1]
        if not path:
            return jsonify({'error': logs[-1], 'logs': logs, 'time': elapsed_time}), 500
        else:
            return jsonify({'error': 'An error occurred while finding path', 'logs': logs, 'time': elapsed_time}), 500
>>>>>>> REPLACE
```

This change will ensure that the elapsed time is always included in the server response, even when an error occurs. This should allow the elapsed time to be displayed on the client side as expected.