aider: Added console.log to debug logsHtml string in client/main.js
# Aider chat conversation:

USER: indeed, this string is not printed.

ASSISTANT: I see. It seems like the static string "This is a test string" is not being displayed in the logs box. This could be due to a number of reasons.

One possibility is that the logs box is not being updated correctly. Another possibility is that there is an error in the JavaScript code that is preventing the logs box from being updated.

To troubleshoot this issue, we could try adding a console.log statement right before the logs box is updated. This would allow us to see if the logsHtml string is being constructed correctly.

Here is the *SEARCH/REPLACE* block to add a console.log statement in `client/main.js`:

client/main.js
```javascript
<<<<<<< SEARCH
        logsHtml += '\n<p>This is a test string</p>';
        logsHtml += '</pre>';
        logsElement.innerHTML = logsHtml;
=======
        logsHtml += '\n<p>This is a test string</p>';
        logsHtml += '</pre>';
        console.log(logsHtml);
        logsElement.innerHTML = logsHtml;
>>>>>>> REPLACE
```

This change will print the logsHtml string to the console right before it is used to update the logs box. If the logsHtml string is correct, then the issue is likely with how the logs box is being updated. If the logsHtml string is not correct, then the issue is likely with how the logsHtml string is being constructed.