aider: Moved elapsed time and discovered pages statistics to a new bullet pointed list in the UI.
# Aider chat conversation:

USER: I decided to move elapsed time and other stats out of the textbox. the layout of the page should be: bulletpointed list of the path (as it is now), textbox with all discovered pages (as it is now) and then a bullet pointed list with various statistics of the algorithm including elapsed time and number of discovered pages.

ASSISTANT: To implement this change, we need to modify `client/index.html` and `client/main.js`.

In `client/index.html`, we need to add a new div for the statistics.

In `client/main.js`, we need to update the fetch request's `.then` block to populate the statistics div with the elapsed time and number of discovered pages.

Here are the *SEARCH/REPLACE* blocks:

client/index.html
```html
<<<<<<< SEARCH
    <div id="logs" style="width: 80%; height: 200px; overflow: auto; border: 1px solid black; margin-top: 20px;"></div>
    <script src="/static/main.js?v=3"></script>
=======
    <div id="logs" style="width: 80%; height: 200px; overflow: auto; border: 1px solid black; margin-top: 20px;"></div>
    <div id="stats" style="width: 80%; margin-top: 20px;"></div>
    <script src="/static/main.js?v=3"></script>
>>>>>>> REPLACE
```

client/main.js
```javascript
<<<<<<< SEARCH
        logsHtml += '\n<p>This is a test string</p>';
        logsHtml += '</pre>';
        console.log(logsHtml);
        logsElement.innerHTML = logsHtml;
=======
        logsHtml += '</pre>';
        logsElement.innerHTML = logsHtml;
        var statsElement = document.getElementById('stats');
        var statsHtml = '<ul>';
        statsHtml += '<li>Elapsed time: ' + data.time + '</li>';
        statsHtml += '<li>Number of discovered pages: ' + data.logs.length + '</li>';
        statsHtml += '</ul>';
        statsElement.innerHTML = statsHtml;
>>>>>>> REPLACE
```

This change will create a new bullet pointed list below the logs box that displays the elapsed time and number of discovered pages.