aider: Added console.log to debug statsHtml output in client/main.js
# Aider chat conversation:

USER: ok, so now we need to output into the stats div. i see the code for this. but there is no output on the webpage.

ASSISTANT: If the statistics are not being displayed on the webpage, it could be due to a number of reasons.

One possibility is that the `statsElement` is not being updated correctly. Another possibility is that there is an error in the JavaScript code that is preventing the `statsElement` from being updated.

To troubleshoot this issue, we could try adding a console.log statement right before the `statsElement` is updated. This would allow us to see if the `statsHtml` string is being constructed correctly.

Here is the *SEARCH/REPLACE* block to add a console.log statement in `client/main.js`:

client/main.js
```javascript
<<<<<<< SEARCH
        statsHtml += '</ul>';
        statsElement.innerHTML = statsHtml;
=======
        statsHtml += '</ul>';
        console.log(statsHtml);
        statsElement.innerHTML = statsHtml;
>>>>>>> REPLACE
```

This change will print the `statsHtml` string to the console right before it is used to update the statistics div. If the `statsHtml` string is correct, then the issue is likely with how the statistics div is being updated. If the `statsHtml` string is not correct, then the issue is likely with how the `statsHtml` string is being constructed.